requirements.txt
apache-airflow==2.7.1
pandas==1.5.3
numpy==1.24.3
yfinance==0.2.36
great-expectations==0.17.19
scikit-learn==1.0.2
pytest==7.4.3
dvc==3.30.1
python-dotenv==1.0.0
scipy==1.10.1
statsmodels==0.13.5
plotly==5.18.0
ta==0.10.2
requests==2.31.0
PyYAML==6.0.1
email-validator>=1.0.5,<2.0.0
